Lesson03. HW03 bug fixes.
1. Fixed getBetweenHalfOpen
2. populateDB.sql - added new meals for User
3. Fixed JdbcMealRepository. Changed "get" method.
4. Tuned spring xml config files to work better with tests.

+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcMealRepository.java
@@ -1,6 +1,7 @@
 package ru.javawebinar.topjava.repository.jdbc;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.support.DataAccessUtils;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
@@ -46,13 +47,8 @@ public class JdbcMealRepository implements MealRepository {
         if (meal.isNew()) {
             Number id = insertMeal.executeAndReturnKey(map);
             meal.setCalories(id.intValue());
-        } else if (namedParameterJdbcTemplate.update(   "UPDATE meals " +
-                                                            "SET    date_time=:date_time," +
-                                                            "       description=:description," +
-                                                            "       calories=:calories " +
-                                                            "WHERE  id =:id" +
-                                                            "       AND user_id=:user_id", map) == 0) {
-
+        } else if (namedParameterJdbcTemplate.update("UPDATE meals SET date_time=:date_time, " +
+                "description=:description, calories=:calories WHERE id =:id AND user_id=:user_id", map) == 0) {
             return null;
         }
         return meal;
@@ -60,33 +56,24 @@ public class JdbcMealRepository implements MealRepository {
 
     @Override
     public boolean delete(int id, int userId) {
-        return jdbcTemplate.update( "DELETE FROM    meals " +
-                                        "WHERE          id = ?" +
-                                        "               AND user_id = ?", id, userId) != 0;
+        return jdbcTemplate.update("DELETE FROM meals WHERE id = ? AND user_id = ?", id, userId) != 0;
     }
 
     @Override
     public Meal get(int id, int userId) {
-        return jdbcTemplate.query(  "SELECT * " +
-                                        "FROM   meals " +
-                                        "WHERE  id = ?" +
-                                        "       AND user_id = ?", ROW_MAPPER, id, userId).stream().findAny().orElse(null);
+        List<Meal> resultList = jdbcTemplate.query("SELECT * FROM meals WHERE id = ? AND user_id = ?", ROW_MAPPER,
+                id, userId);
+        return DataAccessUtils.singleResult(resultList);
     }
 
     @Override
     public List<Meal> getAll(int userId) {
-        return jdbcTemplate.query(  "SELECT     * " +
-                                        "FROM       meals " +
-                                        "WHERE      user_id = ? " +
-                                        "ORDER BY   date_time desc", ROW_MAPPER, userId);
+        return jdbcTemplate.query("SELECT * FROM meals WHERE user_id = ? ORDER BY date_time desc", ROW_MAPPER, userId);
     }
 
     @Override
     public List<Meal> getBetweenHalfOpen(LocalDateTime startDateTime, LocalDateTime endDateTime, int userId) {
-        return jdbcTemplate.query(  "SELECT     * " +
-                                        "FROM       meals " +
-                                        "WHERE      user_id = ? " +
-                                        "           AND date_time BETWEEN ? AND ? " +
-                                        "ORDER BY   date_time desc", ROW_MAPPER, userId, startDateTime, endDateTime);
+        return jdbcTemplate.query("SELECT * FROM meals WHERE user_id = ? AND date_time >= ? AND date_time < ? " +
+                "ORDER BY date_time desc", ROW_MAPPER, userId, startDateTime, endDateTime);
     }
 }
